movement: fix cache usage for random walk point
walkTo(worldPoint, radius) generated a new random world point which it
passed to Walking. This means that the cache would, depending on the
radius size, almost never be hit. Building a world area around the point
with the given radius allows us to use the world area cache and still
get a randomized world point.