item charges: fix tracking explorer ring charges
668d99dfef72c900b1c6c7a5ab121863ddb06252 no longer causes the varbit to
be checked on login due to other varbits changing. If the value is 0,
for full charges, no event will be fired because the default varbit
value is 0.

This initializes the charges to 0 on login to handle this case

(cherry picked from commit c4e32f8aebfddf50fd4abb27d7d7b2732566834c)