api: add widget tick api
The provided callbacks are called when layers and interfaces are ticked.

WidgetOverlay has been modified to use the tick api to position the
widget. The force move api has been removed.

WidgetOverlay has to account for either a subchange listener, running
during packet processing, from changing the widget. Or a script event, run
after the client tick event, changing it. Either are bad, the former
affecting where the widget is when ticked, breaking interacting, and the
latter its position next frame, for rendering.

To account for this it positions the widget prior to the client ticking
it and then again, after the client tick is finished.

Due to the use of the manual layer and renderAfterInterface draw hooks,
this now prevents WidgetOverlays on interfaces which aren't loaded from
rendering each frame.

(cherry picked from commit 2cc5f73c1f66199abc2ac0efaf5820bff64333b8)