loot tracker: Fix chest re-opening detection
Presently, it is possible to have loot from Theatre of Blood not be
captured by the loot tracker by looting the lobby chest for one run and
looting inside of the theatre for the following run, because
`chestLooted` would never be reset. (because all game state changes to
LOADING would be within the theatre, which is instanced) This commit
changes that behavior to track the previous instanced state and reset
`chestLooted` whenever entering or leaving an instanced area.

(cherry picked from commit 02fdca16e62150c1e4eee102ba1ee27139980e87)