loottracker: fix whisperer tracking deaths
The instance is not region-aligned and so there are 4 regions meeting in the middle of it, making this important. Use player location for both sides instead.

Also clear the variables if the delay times out.

(cherry picked from commit 2c61107c9ccbe9a9fcd7dc258c61013c2c4b17e4)