client: add chat icons api
With the client now able to do a full js5 reload without restarting,
this invalidates the chat icons, amonger other things. So they have to
be reloaded once the client chnages to STARTING state.

This adds an api to make this easier, since it involves a bit of
tracking to do correctly, and uses it in the friend notes and emoji
plugins.

Chat commands has to load the pets enum and then dynamically create the
pet images from the item sprites, which is hard to do, and so has been
left using the existing api.

(cherry picked from commit 3145903503f0ad322eba280faf8168e399f41514)