gpu: reinitialize opencl variables on init
This fixes a crash due to a double free of tile heights on the second start of the opencl manager, due to it not zeroing the tile height image in cleanup. Just reinitialize all the fields to be safe.

(cherry picked from commit a7f0bd97aaf24b2071bdf18162426552e4755b78)