object indicators: revert "make hull fill color consistent"
The behavior change seems unpopular. Go back to using a=50 for hull, a/12 for tile and clickbox, and then nothing for outline.

Default color picker to a=50 because determining the in-use fill becomes more complex now that it depends on the type of highlight.

(cherry picked from commit 639de097da53050a1203ad41a50424c205087513)