ui: remove contain in screen ALWAYS
Previously with custom chrome, the window had no title bar and dragging
was supported via the substance root pane using a mouse listener which
called setLocation. Thus restricting the frame to a single screen was
possible with overring setLocation.

Now we have a native window with a native titlebar, dragging is
performed natively, and the location is not set via setLocation(). We
had attempted to work around this by registering a component moved
listener, but it runs after the window is actually moved and the end
result is we end up fighting with the native drag system, which is
visibily jarring.

It looks like it might be possible to do this via a custom wndproc
checking for SC_MOVE, but it would probably require some extensive
flatlaf changes and also is not portable. Remove the option instead.

(cherry picked from commit af737336dc15a210450901b36a4d4b7ee926bd73)