notifier: remove double shell escaping
With the exception of tha osascript invocation, all instances of the message and title are passed to ProcessBuilder as a List, which already handles escaping correctly. osascript requires escaping the strings when interpreted by AppleScript, and not due to the shell.

(cherry picked from commit 54f206647da17ae5e1c9944125d719b07bea27df)