timers: Remove tzhaar timer defeat message trigger
It is not clear what purpose this trigger serves given that there is
also a GameStateChanged trigger which accomplishes the same thing.
Because the chat message and game state fire when dying in the fight
caves or inferno, the resulting `config.startTime(null)` call after the
timer is removed leads to an ElapsedTimer being created with a null
start time, which causes an NPE during infobox rendering. Because this
trigger is redundant, it can safely be removed, which also prevents the
aforementioned null start time ElapsedTimer from being created.

(cherry picked from commit 3628c48942d1bdcb4062965d07cf03267af7365e)