menu swapper: fix crash performing Use swap on items with submenus
This was calling swap("use") on a submenu menu entry, findIndex() was
finding the "Use" from the top level menu, calling swap() which then was
passing the submenu MenuEntry to client.setMenuEntries()

optionIndexes used to be heavily used, but now it is only used for built
in swaps and for item "Use" swap. Change it to be lazily computed and be
per-menu.

Pass the Menu around to the various functions and use it instead of
assuming the top level menu.

(cherry picked from commit c36527d0f9c9ccef79fb900ec7d4a7c4af8f63ee)