wiki: fix view dps button to use absolute_center y position mode
In the bank the Set/Stat bonus layer uses absolute_center so that it
gets repositioned when the bank is resized. Do the same for the View dps
button so it also gets repositioned when the bank is resized.

The equipment screen uses absolute_top instead, so pull the position
mode from the set bonus parent.

(cherry picked from commit 9070e8711087d43a04c4c8913bf8e78cf09a1903)